sillycat-authcenter-python/requirements.txt
fastapi==0.61.2
gunicorn==20.0.4
uvicorn==0.12.2
uvloop==0.14.0
httptools==0.1.1
PyJWT==2.1.0
pydantic==1.7.4
psycopg2-binary==2.8.6
cryptography==3.4.7